move more classes from control
