+++ client/ayon_core/tools/publisher/models/__init__.py
@@ -1,9 +1,15 @@
 from .create import CreatorItem
-from .publish import PublishReportMaker
+from .publish import (
+    PublishReportMaker,
+    PublishValidationErrors,
+    PublishPluginsProxy,
+)
 
 
 __all__ = (
     "CreatorItem",
 
     "PublishReportMaker",
+    "PublishValidationErrors",
+    "PublishPluginsProxy",
 )
